% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attach_to_list}
\alias{attach_to_list}
\title{Combine two lists}
\usage{
attach_to_list(base_list, extra_list)
}
\arguments{
\item{base_list}{List to attach objects to.}

\item{extra_list}{List from which elements should be attached to \code{base_list}.
Duplicated objects are overwritten.}
}
\value{
List.
}
\description{
Combine two lists
}
