% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{in_range}
\alias{in_range}
\title{Check if value fits to a range}
\usage{
in_range(x, range)
}
\arguments{
\item{x}{Numeric value.}

\item{range}{Vector of length 2, storing range change limits.}
}
\value{
A logical vector indicating which elements of \code{x} fit into the specified \code{range}.
}
\description{
Check if value fits to a range
}
