% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lget_attr}
\alias{lget_attr}
\title{Extract attribute of each element from a set of lists}
\usage{
lget_attr(list_obj, attribute)
}
\arguments{
\item{list_obj}{List of lists. Each nested list should contain \code{el_name} object.}

\item{attribute}{Name of the attribute to extract from each object.}
}
\value{
Vector of the same length, storing extracted attributes.
}
\description{
Extract attribute of each element from a set of lists
}
