% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBuilder.R
\name{lookup}
\alias{lookup}
\title{lookup}
\usage{
lookup(id, operator, value)
}
\arguments{
\item{id}{data frame column id}

\item{operator}{filter operator as defined within queryBuilder}

\item{value}{filter value}
}
\value{
string representation of a single filter
}
\description{
internal function to create a filter condition based on id, operator and value
}
