% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.R
\docType{data}
\name{operator_name_prefix}
\alias{operator_name_prefix}
\title{Prefix used for renaming operators names}
\format{
An object of class \code{character} of length 1.
}
\usage{
operator_name_prefix
}
\description{
Required due to erroneous operations on objects with names such as 'in' or 'for'.
}
\keyword{datasets}
