% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.R
\name{prefix_operators_name}
\alias{prefix_operators_name}
\title{Rename operators with the provided prefix}
\usage{
prefix_operators_name(operators)
}
\arguments{
\item{operators}{List storing \link{queryOperator}s.}
}
\description{
Rename operators with the provided prefix
}
