% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.R
\name{query_to_expr_bare}
\alias{query_to_expr_bare}
\title{Convert query definition to expression}
\usage{
query_to_expr_bare(query, operators, conditions, keep_na)
}
\arguments{
\item{query}{Query definition (see \link{queryRule} and \link{queryGroup}).}

\item{operators}{List storing \link{queryOperator}s.}

\item{conditions}{List storing \link{queryCondition}s.}

\item{keep_na}{Should each rule expression be extended with rule excluding/including 'NA' values?}
}
\value{
Character value storing expression to be parsed.
}
\description{
Convert query definition to expression
}
