% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBuilder.R
\name{queryBuilder}
\alias{queryBuilder}
\title{queryBuilder}
\usage{
queryBuilder(data = NULL, autoassign = FALSE, filters = list(),
  default_condition = "AND", allow_empty = FALSE, display_errors = TRUE,
  display_empty_filter = TRUE, width = NULL, height = NULL)
}
\arguments{
\item{data}{data frame}

\item{filters}{list of lists containing filter parameters}
}
\description{
create a queryBuilder widget from a series of filters
}

