% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{queryBuilderConfig}
\alias{queryBuilderConfig}
\title{Default object storing 'queryBuilder' configuration.}
\format{
An object of class \code{queryBuilderConfig} (inherits from \code{R6}) of length 8.
}
\usage{
queryBuilderConfig
}
\description{
Default object storing 'queryBuilder' configuration.
}
\keyword{datasets}
