% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBuilder.R
\name{recurseFilter}
\alias{recurseFilter}
\title{recurseFilter}
\usage{
recurseFilter(filter = NULL)
}
\arguments{
\item{filter}{filters output from queryBuilder htmlWidget}
}
\value{
string representation of all filters combined
}
\description{
internal recursive function to process filter
}

