% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_by_name}
\alias{remove_by_name}
\title{Remove list elements by their names}
\usage{
remove_by_name(list_obj, ids)
}
\arguments{
\item{list_obj}{List object.}

\item{ids}{Objects names to be removed.}
}
\value{
List.
}
\description{
Remove list elements by their names
}
