% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.R
\name{rule_to_expr}
\alias{rule_to_expr}
\title{Convert rule definition to expression}
\usage{
rule_to_expr(rule, operators, keep_na = FALSE)
}
\arguments{
\item{rule}{Rule definition (see \link{queryRule}).}

\item{operators}{List storing \link{queryOperator}s.}

\item{keep_na}{Should the expression be extended with rule excluding/including 'NA' values?}
}
\value{
Character value storing expression to be parsed.
}
\description{
Convert rule definition to expression
}
