% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substitute_q}
\alias{substitute_q}
\title{Substitute expression stored as a variable}
\usage{
substitute_q(x, env)
}
\arguments{
\item{x}{Expression to be substituted.}

\item{env}{List of arguments to substitute for x.}
}
\description{
Substitute expression stored as a variable
}
