% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{err_msg}
\alias{err_msg}
\title{Generate error message}
\usage{
err_msg(msg, ..., .envir = parent.frame())
}
\arguments{
\item{msg}{Character string interpreted by \link[glue]{glue}.}

\item{...}{Extra arguments passed to \link[glue]{glue}.}

\item{.envir}{Environment to evaluate each expression in, passed to \link[glue]{glue}.}
}
\value{
Executed error with interpolated message.
}
\description{
Generate error message
}
